Rename project for better naming of things
