+++ ShiftsLogger.tonyissa/ShiftsLogger.Client/App.config
