--- ShiftsLogger.tonyissa/ShiftsLogger.UI/Program.cs
@@ -1,1 +1,0 @@
-﻿
