+++ ShiftsLogger.tonyissa/ShiftsLogger.tonyissa.sln
@@ -3,12 +3,14 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.11.35303.130
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ShiftsLogger.API", "ShiftsLogger.API\ShiftsLogger.API.csproj", "{FE340480-EDD4-4089-9C59-6577E7DBDF13}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ShiftsLogger.API", "ShiftsLogger.API\ShiftsLogger.API.csproj", "{FE340480-EDD4-4089-9C59-6577E7DBDF13}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{561A78C2-690A-496E-9E80-E3FBEE45FB16}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ShiftsLogger.Client", "ShiftsLogger.Client\ShiftsLogger.Client.csproj", "{09779334-82E2-46B1-9832-748450BA4E46}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -20,6 +22,10 @@ Global
 		{FE340480-EDD4-4089-9C59-6577E7DBDF13}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FE340480-EDD4-4089-9C59-6577E7DBDF13}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FE340480-EDD4-4089-9C59-6577E7DBDF13}.Release|Any CPU.Build.0 = Release|Any CPU
+		{09779334-82E2-46B1-9832-748450BA4E46}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{09779334-82E2-46B1-9832-748450BA4E46}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{09779334-82E2-46B1-9832-748450BA4E46}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{09779334-82E2-46B1-9832-748450BA4E46}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
